requirements.txt
pymysql
pandas
numpy
google-cloud-vision

spacy

# common ML functions and logging
boto3
git+https://gitlab.com/rauner/buycycle.git@main#egg=buycycle